+++ reusables/__init__.py
@@ -5,6 +5,8 @@ from .dt import *
 from .log import *
 from .browser import *
 from .wrappers import *
+from .multiprocess_helpers import *
+from .web import *
 
 __author__ = "Chris Griffith"
-__version__ = "0.5.2"
+__version__ = "0.6.0"